Revert "Revert "lib/envflag: apply -secret.flags inside envflag.Parse function (#9942)""
This reverts commit c90c7c31234b7677da751ed51f9760fcccc6c805.
